breaking change: refactor to support Node type-stripping
Node 23.6.0 support TypeScript import out of the box. There is a restriction however: only erasable syntax is support. This means that type imports must explicitly be marked as such, and enums are not supported. This commit refactors the lib to support it.

As part of this work, many issues were found and solved. See individual changelog files for details.

This commit also updates ESLint to its latest version.